Rename instantiate module to resolve
