+++ base/src/lib.rs
@@ -14,10 +14,10 @@ pub mod ast;
 pub mod error;
 pub mod fixed;
 pub mod fnv;
-pub mod instantiate;
 pub mod kind;
 pub mod metadata;
 pub mod pos;
+pub mod resolve;
 pub mod scoped_map;
 pub mod source;
 pub mod symbol;
