+++ base/src/resolve.rs
